+++ src/main/resources/application.properties
@@ -21,7 +21,7 @@ spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 # Hibernate ddl-auto - Use 'update' for development so tables are created automatically
 spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
+spring.jpa.show-sql=false
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
 
 vaadin.frontend.hotdeploy=true
